fix: split routes and controller for react-migration
